Enable partial windows when generating BP passports
This allows us to not drop passports when page count doesn't satisfy the number of passports to be generated.

For example, if you are trying to generate 10 passports with 1 row and 2 column. You are only able to generate 8 passports with 4 passports in each PDF. Instead this allows us to create an extra PDF with the remaining 2 passports in it.