Generate max 100 passport pages per PDF (#25)
* Enable partial windows when generating BP passports

This allows us to not drop passports when page count doesn't satisfy the number of passports to be generated.

For example, if you are trying to generate 10 passports with 1 row and 2 column. You are only able to generate 8 passports with 4 passports in each PDF. Instead this allows us to create an extra PDF with the remaining 2 passports in it.

* Add `pageCount` as param for `PassportsGenerator`

We want every PDF to contain max of 100 BP Passport pages.  So, if user want to generate less than 100 passports. They all will be in a single PDF.

* Remove code to read page count from template file

* Rename approved PDF images in tests

Now that we are generating 100 passport pages per PDF. All of the passports in tests will be in single page.